# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the official data model for BARAFIRI API documentation.",
        "title": "Barafiri Backend API documentation",
        "termsOfService": "http://my-gpi.io/terms/",
        "contact": {
            "name": "Endy Apinageri",
            "url": "https://www.endyapina.com",
            "email": "endy.apinageri@my-gpi.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8112",
    "basePath": "/api/v1",
    "paths": {
...
        "/business/": {
            "post": {
                "description": "Send user data to the main authentication endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Endpoint to send user data for registration",
                "parameters": [
                    {
                        "description": "The Business Registration Data",
                        "name": "BusinessRegistration",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.BusinessRegistration"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data ",
                        "schema": {
                            "$ref": "#/definitions/controller.BusinessProfile"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            },
            "delete": {
                "description": "Perform Delete endpoint for Business Registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Endpoint to delete a user business information",
                "parameters": [
                    {
                        "description": "The Business Registration Data",
                        "name": "BusinessProfile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.BusinessProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data ",
                        "schema": {
                            "$ref": "#/definitions/controller.BusinessProfile"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/business/all": {
            "get": {
                "description": "Get an array of all business and business data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Endpoint to get all businesses on BARAFIRI",
                "responses": {
                    "200": {
                        "description": "Success data ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.BusinessResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/business/update": {
            "post": {
                "description": "Perform Update endpoint for Business Registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Endpoint to update user business data",
                "parameters": [
                    {
                        "description": "The Business Registration Data",
                        "name": "BusinessProfile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.BusinessProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data ",
                        "schema": {
                            "$ref": "#/definitions/controller.BusinessProfile"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/business/{buid}": {
            "get": {
                "description": "Get an array of all business and business data by business unique ids",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Endpoint to get business information by id",
                "parameters": [
                    {
                        "description": "The Business unique id string",
                        "name": "businessid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data ",
                        "schema": {
                            "$ref": "#/definitions/controller.BusinessResponse"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
...
        "/public/business/products/all": {
            "get": {
                "description": "Get business information and products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public API"
                ],
                "summary": "Endpoint to get business information based on business unique id",
                "responses": {
                    "200": {
                        "description": "Success data. Check response body for data containing business data and all product array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.BusinessResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/public/product/all": {
            "get": {
                "description": "Get an array of all business and business data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public API"
                ],
                "summary": "Endpoint to get all product information on BARAFIRI",
                "responses": {
                    "200": {
                        "description": "Success data. Check response body for data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.ProductResponseObject"
                            }
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/public/product/{puid}": {
            "get": {
                "description": "Get product by product id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public API"
                ],
                "summary": "Endpoint to get all product information on BARAFIRI",
                "responses": {
                    "200": {
                        "description": "Success data. Check response body for data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.ProductResponseObject"
                            }
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.BusinessProfile": {
            "type": "object",
            "properties": {
                "business_address": {
                    "type": "string"
                },
                "business_description": {
                    "type": "string"
                },
                "business_id": {
                    "type": "string"
                },
                "business_name": {
                    "type": "string"
                },
                "business_type": {
                    "type": "string"
                },
                "cac_registration_number": {
                    "type": "string"
                },
                "core_business_functions": {
                    "type": "string"
                },
                "core_business_training": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "industry": {
                    "type": "string"
                },
                "operational_support": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "years_of_operation": {
                    "type": "integer"
                }
            }
        },
        "controller.BusinessRegistration": {
            "type": "object",
            "properties": {
                "business_address": {
                    "type": "string"
                },
                "business_description": {
                    "type": "string"
                },
                "business_name": {
                    "type": "string"
                },
                "business_type": {
                    "type": "string"
                },
                "cac_registration_number": {
                    "type": "string"
                },
                "core_business_functions": {
                    "type": "string"
                },
                "core_business_training": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "operational_support": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "years_of_operation": {
                    "type": "integer"
                }
            }
        },
        "controller.BusinessResponse": {
            "type": "object",
            "properties": {
                "business_address": {
                    "type": "string"
                },
                "business_description": {
                    "type": "string"
                },
                "business_id": {
                    "type": "string"
                },
                "business_name": {
                    "type": "string"
                },
                "business_type": {
                    "type": "string"
                },
                "cac_registration_number": {
                    "type": "string"
                },
                "core_business_functions": {
                    "type": "string"
                },
                "core_business_training": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "operational_support": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.ProductResponseObject"
                    }
                },
                "website": {
                    "type": "string"
                },
                "years_of_operation": {
                    "type": "integer"
                }
            }
        },
...
        "controller.ProductDiscount": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "discountPercentage": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "controller.ProductImage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "controller.ProductResponseObject": {
            "type": "object",
            "properties": {
                "discount": {
                    "$ref": "#/definitions/controller.ProductDiscount"
                },
                "industry": {
                    "type": "string"
                },
                "item_left": {
                    "type": "integer"
                },
                "price": {
                    "type": "string"
                },
                "product_description": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "product_images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.ProductImage"
                    }
                },
                "product_name": {
                    "type": "string"
                },
                "ratings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.Rating"
                    }
                },
                "section": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.Tags"
                    }
                }
            }
        },
...
        "controller.Rating": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_comments": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_rating": {
                    "type": "integer"
                }
            }
        },
...
        "controller.ResponseObject": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object"
                },
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controller.Tags": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}